Use Fair STM QDisc; improve simulation output
